Fix followup email selection
The changes introduced in #125 caused us to stop sending email
notifications to admins, because they don't show up as "collaborators"
on an app.

There's actually already code to find all relevant users for a given
email, so this updates the NotificationUpdater to use it. This also
addresses an apparent issue that NotificationUpdater assumed all
messages were "app messages", but it's also possible to send messages
directly to users. The mediator used accounts for this.

Finally, the code and tests also now consistently handle the app owner
receiving notifications. It's not clear if owners were not receiving
notifications in production, but the tests were inconsistent here.